Moving to group middleware into related modules, following the pattern established in prison staff hub
