+++ server/middleware/setupRequestParsing.ts
@@ -1,0 +1,8 @@
+import express, { Router } from 'express'
+
+export default function setUpWebRequestParsing(): Router {
+  const router = express.Router()
+  router.use(express.json())
+  router.use(express.urlencoded({ extended: true }))
+  return router
+}
