+++ server/utils/nunjucksSetup.ts
@@ -1,8 +1,28 @@
+/* eslint-disable no-param-reassign */
 import nunjucks from 'nunjucks'
 import express from 'express'
 import * as pathModule from 'path'
 
-export default function nunjucksSetup(app: express.Application, path: pathModule.PlatformPath): void {
+const production = process.env.NODE_ENV === 'production'
+
+export default function nunjucksSetup(app: express.Express, path: pathModule.PlatformPath): void {
+  app.set('view engine', 'njk')
+
+  app.locals.asset_path = '/assets/'
+  app.locals.applicationName = 'HMPPS Typescript Template'
+
+  // Cachebusting version string
+  if (production) {
+    // Version only changes on reboot
+    app.locals.version = Date.now().toString()
+  } else {
+    // Version changes every request
+    app.use((req, res, next) => {
+      res.locals.version = Date.now().toString()
+      return next()
+    })
+  }
+
   const njkEnv = nunjucks.configure(
     [
       path.join(__dirname, '../../server/views'),
